+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4.js
@@ -9,4 +9,5 @@
 //= require ./analytics-ga4/ga4-ecommerce-tracker
 //= require ./analytics-ga4/ga4-form-tracker
 //= require ./analytics-ga4/ga4-auto-tracker
+//= require ./analytics-ga4/ga4-smart-answer-results-tracker
 //= require ./analytics-ga4/init-ga4