+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-smart-answer-results-tracker.js
@@ -1,0 +1,55 @@
+window.GOVUK = window.GOVUK || {}
+window.GOVUK.Modules = window.GOVUK.Modules || {}
+window.GOVUK.analyticsGa4 = window.GOVUK.analyticsGa4 || {};
+
+(function (Modules) {
+  'use strict'
+
+  function Ga4SmartAnswerResultsTracker (module) {
+    this.module = module
+  }
+
+  Ga4SmartAnswerResultsTracker.prototype.init = function () {
+    var consentCookie = window.GOVUK.getConsentCookie()
+
+    if (consentCookie && consentCookie.settings) {
+      this.startModule()
+    } else {
+      this.startModule = this.startModule.bind(this)
+      window.addEventListener('cookie-consent', this.startModule)
+    }
+  }
+
+  // triggered by cookie-consent event, which happens when users consent to cookies
+  Ga4SmartAnswerResultsTracker.prototype.startModule = function () {
+    if (window.dataLayer) {
+      this.trackResults()
+      this.module.addEventListener('click', this.handleClick.bind(this))
+    }
+  }
+
+  Ga4SmartAnswerResultsTracker.prototype.trackResults = function () {
+    var ecommerceSchema = window.GOVUK.analyticsGa4.core.ecommerceHelperFunctions.populateEcommerceSchema({
+      element: this.module,
+      resultsId: 'ga4-ecommerce-result-count'
+    })
+
+    window.GOVUK.analyticsGa4.core.ecommerceHelperFunctions.clearEcommerceObject()
+    window.GOVUK.analyticsGa4.core.sendData(ecommerceSchema)
+  }
+
+  Ga4SmartAnswerResultsTracker.prototype.handleClick = function (event) {
+    if (event.target.getAttribute('data-ga4-ecommerce-path')) {
+      var ecommerceSchema = window.GOVUK.analyticsGa4.core.ecommerceHelperFunctions.populateEcommerceSchema({
+        element: this.module,
+        resultsId: 'ga4-ecommerce-result-count',
+        event: event
+      })
+
+      window.GOVUK.analyticsGa4.core.ecommerceHelperFunctions.clearEcommerceObject()
+      window.GOVUK.analyticsGa4.core.sendData(ecommerceSchema)
+    }
+  }
+
+  Modules.Ga4SmartAnswerResultsTracker = Ga4SmartAnswerResultsTracker
+})(window.GOVUK.Modules)